+++ flutter_arc_speed_dial/example/lib/main.dart
@@ -94,6 +94,17 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   Widget _getBodyWidget() {
-    return Container();
+    return Container(
+//      alignment: Alignment.bottomRight,
+//      child: SizedBox(
+//        width: 300,
+//        height: 200,
+//        child: RaisedButton(
+//            onPressed: () {
+//              print('success');
+//            },
+//            child: Text('testing ')),
+//      ),
+    );
   }
 }