+++ flutter_arc_speed_dial/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.3] - 20200621
+* Fixed invisible button still able to click issue
+* Added example project
+
 ## [1.0.2] - 20191201
 * Added two fields (isShowSpeedDial, updateSpeedDialStatus) for manually change menu open or close status and listen if any changes within the widget